+++ src/System.Text.Json/src/System.Text.Json.csproj
@@ -23,7 +23,8 @@
     <Compile Include="System\Text\Json\JsonHelpers.cs" />
     <Compile Include="System\Text\Json\JsonHelpers.Date.cs" />
     <Compile Include="System\Text\Json\JsonTokenType.cs" />
-    <Compile Include="System\Text\Json\Node\JsonNode.Parse.cs" />
+    <Compile Include="System\Text\Json\Node\JsonNode.ParsingHelpers.cs" />
+    <Compile Include="System\Text\Json\Node\JsonNode.Parsing.cs" />
     <Compile Include="System\Text\Json\Node\JsonObjectProperty.cs" />
     <Compile Include="System\Text\Json\Node\JsonNull.cs" />
     <Compile Include="System\Text\Json\ThrowHelper.cs" />
